refactor(harness)!: consolidate each harness to _<harness>_sync.py + _<harness>_turn.py
Folds the pydantic-ai/langgraph _async + _langgraph_messages helpers into their
turn/sync modules (stream_*_events -> _<h>_turn.py, emit_langgraph_messages ->
_langgraph_sync.py); public facade names are unchanged. Relocates the OpenAI harness
Turn + convert_openai_to_agentex_events tap into _modules/_openai_turn.py /
_modules/_openai_sync.py, leaving back-compat shims at providers/_modules/{openai_turn,
sync_provider}.py so the adk.providers namespace + CLI template keep working (the larger
openai.py Temporal/MCP provider stays under adk.providers). Merges the duplicate
_sync / _sync_unified test modules into one per harness.
